+++ src/lang/builtins_harbor.ml
@@ -31,7 +31,7 @@ end
 
 module Make (Harbor : T) = struct
   let name_up = String.uppercase_ascii Harbor.name
-  let resp_t = Lang.string_getter_t ()
+  let resp_t = Lang.getter_t Lang.string_t
   let () = Lang.add_module ("harbor." ^ Harbor.name)
 
   let () =